backend/requirements.txt
web3
Flask
flask
Flask-session
cryptography
qrcode
opencv-python
flask-cors
ipfshttpclient
celery
redis
requests
dotenv
pillow
pymongo
pandas
bcrypt
easyocr
datetime
qrcode
numpy
pyzbar
torch
pytesseract 
opencv-python
